Update client.py
The previous version, opening a file and then not closing it is not a good idea and might cause unwanted file locks or, the worst scenario - leaks. Especially if file size is large.   

Besides this, you might want to move `requests.post` piece into the `with open` block - thus, you'll be able to send the file without reading into memory.